backend/requirements.txt
flask
flask-cors
boto3
pytesseract
pdf2image
Pillow
PyMuPDF
langchain
faiss-cpu